requirements.txt
Flask==0.12.4
Werkzeug<0.13.0,>=0.12.0
GoogleAppEngineCloudStorageClient==1.9.22.1
requests
